# ranking: Optimize dominating query